pyproject.toml
[project]
name = "doc-analyzer"
version = "0.1.0"
description = "LLM-powered document quality analyzer for proposals, kickoffs, and call transcripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.150.0",
    "slack-sdk>=3.33.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
doc-analyzer = "doc_analyzer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
